requirements.txt
annotated-types==0.7.0
anyio==4.9.0
Authlib==1.3.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cryptography==45.0.3
cycler==0.12.1
deprecation==2.1.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fonttools==4.58.4
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.71.0
grpcio-health-checking==1.71.0
grpcio-tools==1.71.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.8
protobuf==5.29.4
psutil==7.0.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
seaborn==0.13.2
sentry-sdk==2.31.0
setproctitle==1.3.6
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.5.0
validators==0.34.0
wandb==0.20.1
weaviate-client==4.14.4
